Refactor: rename project/tests
