+++ vamp-test/.project
@@ -1,8 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>database-prototype-app</name>
+	<name>vamp-test</name>
 	<comment></comment>
 	<projects>
+		<project>vamp</project>
 	</projects>
 	<buildSpec>
 		<buildCommand>
