+++ vamp-test/proguard-project.txt
