+++ vamp-test/project.properties
