+++ vamp-test/res/drawable-hdpi/ic_launcher.png
