+++ vamp-test/res/drawable-ldpi/ic_launcher.png
