+++ vamp-test/res/drawable-mdpi/ic_launcher.png
