+++ vamp-test/res/drawable-xhdpi/ic_launcher.png
