+++ vamp-test/res/values/strings.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <string name="app_name">Database Prototype Tests</string>
+    <string name="app_name">Android Music Player tests</string>
 
 </resources>
