+++ vamp/.classpath
