+++ vamp/proguard-project.txt
