+++ vamp/project.properties
