+++ vamp/res/drawable-ldpi/ic_launcher.png
