+++ vamp/res/layout/main.xml
