+++ vamp/res/values/strings.xml
@@ -2,6 +2,6 @@
 <resources>
 
     <string name="hello">Hello World</string>
-    <string name="app_name">Database Prototype</string>
+    <string name="app_name">Android Music Player</string>
 
 </resources>
